# Add (*Buffer).read1 for 1bit reading