(*Buffer).read(1) is a hot spot and called mainly by readVlcUint when decoding a video.

This change adds an optimized reading function for reading one bit. This reduces time of decoding video by 7-8%.

```
goos: darwin
goarch: arm64
pkg: github.com/gen2brain/mpeg
               │   old.txt   │              new.txt               │
               │   sec/op    │   sec/op     vs base               │
DecodeVideo-12   86.42µ ± 1%   79.38µ ± 1%  -8.14% (p=0.000 n=10)
DecodeAudio-12   48.00µ ± 2%   47.45µ ± 3%       ~ (p=0.190 n=10)
RGBA-12          40.86µ ± 1%   41.80µ ± 2%  +2.31% (p=0.001 n=10)
geomean          55.34µ        54.00µ       -2.42%

               │  old.txt   │              new.txt              │
               │    B/op    │    B/op     vs base               │
DecodeVideo-12   84.00 ± 1%   78.00 ± 1%  -7.14% (p=0.000 n=10)
DecodeAudio-12   47.00 ± 2%   46.00 ± 2%       ~ (p=0.191 n=10)
RGBA-12          40.00 ± 2%   41.00 ± 2%  +2.50% (p=0.009 n=10)
geomean          54.05        52.79       -2.34%

               │   old.txt    │               new.txt               │
               │  allocs/op   │ allocs/op   vs base                 │
DecodeVideo-12   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecodeAudio-12   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
RGBA-12          0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                     ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```