@hajimehoshi 2024-04-12
FYI: I found this hotspot by decoding a video on browsers (https://github.com/hajimehoshi/ebiten/tree/main/examples/video):

![image](https://github.com/gen2brain/mpeg/assets/16950/9a9a8fb0-9169-425c-af4a-4caa546c8d02)